requirements.txt
black[jupyter]
pre-commit
python-dotenv
numpy
boto3>=1.28.58
s3fs==2024.2.0 
pandas==2.1.2
transformers==4.40.1
huggingface_hub==0.20.3
torch==2.2.0
sentencepiece
wandb==0.16.4
accelerate==0.30.0
datasets==2.14.4
evaluate==0.4.1
scikit-learn==1.4.2
bitsandbytes==0.42.0
quanto==0.1.0
scikit-multilearn==0.2.0